refactor: normalize imports under electron folder (#1564)
Fix some import using relative paths.
Move a type file to the types folder
