+++ src/electron/adapters/electron-app-data-adapter.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-
+import { AppDataAdapter } from 'common/browser-adapters/app-data-adapter';
 import { remote } from 'electron';
-import { AppDataAdapter } from '../../common/browser-adapters/app-data-adapter';
 
 export class ElectronAppDataAdapter implements AppDataAdapter {
     public getVersion(): string {
