+++ src/electron/platform/android/scan-results-to-unified-results.ts
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-
 import { InstancePropertyBag, InstanceResultStatus, UnifiedResult } from 'common/types/store-data/unified-data-interface';
 import { UUIDGeneratorType } from 'common/uid-generator';
-import { DictionaryStringTo } from '../../../types/common-types';
+import { DictionaryStringTo } from 'types/common-types';
+
 import { RuleInformation } from './rule-information';
 import { RuleInformationProviderType } from './rule-information-provider-type';
 import { RuleResultsData, ScanResults, ViewElementData } from './scan-results';
