+++ src/electron/views/device-connect-view/components/device-connect-connected-device.tsx
@@ -1,16 +1,17 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { NamedFC } from 'common/react/named-fc';
 import { Icon } from 'office-ui-fabric-react/lib/Icon';
 import { Spinner, SpinnerSize } from 'office-ui-fabric-react/lib/Spinner';
 import * as React from 'react';
-import { NamedFC } from '../../../../common/react/named-fc';
+
+import { DeviceConnectState } from '../../../flux/types/device-connect-state';
 import {
     connectionErrorIcon,
     deviceConnectConnectedDevice,
     deviceConnectSpinner,
     scannedText,
 } from './device-connect-connected-device.scss';
-import { DeviceConnectState } from './device-connect-state';
 
 export interface DeviceConnectConnectedDeviceProps {
     deviceConnectState: DeviceConnectState;
