+++ src/electron/views/device-connect-view/components/device-connect-footer.tsx
@@ -1,10 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { NamedFC } from 'common/react/named-fc';
 import { WindowFrameActionCreator } from 'electron/flux/action-creator/window-frame-action-creator';
 import { WindowStateActionCreator } from 'electron/flux/action-creator/window-state-action-creator';
 import { DefaultButton, PrimaryButton } from 'office-ui-fabric-react/lib/Button';
 import * as React from 'react';
-import { NamedFC } from '../../../../common/react/named-fc';
+
 import { deviceConnectFooter, footerButtonCancel, footerButtonStart } from './device-connect-footer.scss';
 
 export interface DeviceConnectFooterDeps {
