+++ src/electron/views/device-connect-view/components/device-connect-header.tsx
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { NamedFC } from 'common/react/named-fc';
 import * as React from 'react';
 
-import { NamedFC } from '../../../../common/react/named-fc';
 import { deviceConnectHeader } from './device-connect-header.scss';
 import { ElectronLink } from './electron-link';
