+++ src/electron/views/device-connect-view/components/device-connect-port-entry.tsx
@@ -1,13 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { KeyCodeConstants } from 'common/constants/keycode-constants';
 import { DefaultButton, PrimaryButton } from 'office-ui-fabric-react/lib/Button';
 import { MaskedTextField } from 'office-ui-fabric-react/lib/TextField';
 import * as React from 'react';
 
-import { KeyCodeConstants } from 'common/constants/keycode-constants';
 import { DeviceConnectActionCreator } from '../../../flux/action-creator/device-connect-action-creator';
+import { DeviceConnectState } from '../../../flux/types/device-connect-state';
 import { deviceConnectPortEntry, portNumberField } from './device-connect-port-entry.scss';
-import { DeviceConnectState } from './device-connect-state';
 
 export type DeviceConnectPortEntryViewState = {
     deviceConnectState: DeviceConnectState;
