+++ src/electron/views/device-connect-view/components/electron-external-link.tsx
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { NamedFC } from 'common/react/named-fc';
 import { Shell } from 'electron';
 import { Link } from 'office-ui-fabric-react/lib/Link';
 import * as React from 'react';
-import { NamedFC } from '../../../../common/react/named-fc';
 
 export interface ElectronExternalLinkProps {
     href: string;
