+++ src/electron/views/device-connect-view/components/electron-link.tsx
@@ -1,8 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { NamedFC } from 'common/react/named-fc';
 import { shell } from 'electron';
 import * as React from 'react';
-import { NamedFC } from '../../../../common/react/named-fc';
+
 import { ElectronExternalLink } from './electron-external-link';
 
 export type ElectronLinkProps = {
