+++ src/electron/views/device-connect-view/send-app-initialized-telemetry.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { TelemetryEventHandler } from '../../../background/telemetry/telemetry-event-handler';
-import { TelemetryEventSource, TriggeredByNotApplicable } from '../../../common/extension-telemetry-events';
-import { APP_INITIALIZED } from '../../common/electron-telemetry-events';
+import { TelemetryEventHandler } from 'background/telemetry/telemetry-event-handler';
+import { TelemetryEventSource, TriggeredByNotApplicable } from 'common/extension-telemetry-events';
+import { APP_INITIALIZED } from 'electron/common/electron-telemetry-events';
 
 export const sendAppInitializedTelemetryEvent = (telemetryEventHandler: TelemetryEventHandler) => {
     telemetryEventHandler.publishTelemetry(APP_INITIALIZED, {
