+++ src/tests/unit/tests/electron/views/device-connect-view/components/device-connect-connected-device.test.tsx
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+import { DeviceConnectState } from 'electron/flux/types/device-connect-state';
 import {
     DeviceConnectConnectedDevice,
     DeviceConnectConnectedDeviceProps,
 } from 'electron/views/device-connect-view/components/device-connect-connected-device';
-import { DeviceConnectState } from 'electron/views/device-connect-view/components/device-connect-state';
 import { shallow } from 'enzyme';
 import * as React from 'react';
