+++ src/tests/unit/tests/electron/views/device-connect-view/components/device-connect-port-entry.test.tsx
@@ -3,12 +3,12 @@
 import { KeyCodeConstants } from 'common/constants/keycode-constants';
 import { EnumHelper } from 'common/enum-helper';
 import { DeviceConnectActionCreator } from 'electron/flux/action-creator/device-connect-action-creator';
+import { DeviceConnectState } from 'electron/flux/types/device-connect-state';
 import {
     DeviceConnectPortEntry,
     DeviceConnectPortEntryProps,
 } from 'electron/views/device-connect-view/components/device-connect-port-entry';
 import { portNumberField } from 'electron/views/device-connect-view/components/device-connect-port-entry.scss';
-import { DeviceConnectState } from 'electron/views/device-connect-view/components/device-connect-state';
 import { shallow } from 'enzyme';
 import { Button } from 'office-ui-fabric-react/lib/Button';
 import { MaskedTextField } from 'office-ui-fabric-react/lib/TextField';
